requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
torch>=2.4.0
torchvision>=0.19.0
pillow>=10.2.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
numpy>=1.26.3
python-dotenv>=1.0.1
